Reorganize the test files.
